Starting Decode Testing
Appears to be working (in the sense that the decoder is correctly
decoding what is given to it). What is wrong right now: no actual key
put into the file (doing this all in the same function that I create the
encoding in), and the file does not appear to be written correctly.