Starting Canonical Coding
Beginning to setup the header of the file that will have the huffman
tree encoding, starting point is getting the table in Canonical format.
In this commit, I got the encoderTable to order itself in ascending
order by keyLength, with entries that share a keyLength being sorted in
ascending alphabetical order. Added a function to do the sorting of
keyLength (table is set up to already be in alphabetical order), and
updated EncoderEntry to have a character field. Also updated the
initialization from {0} to an EncoderEntry with keyLength=255 so that
any character that doesn't appear will always be at the end.