Reduced EncoderTable to numUniqueCharacter Entries
Had it hardcoded at 256 before, but there really is no need for entries
that aren't in the input string. It was helpful when directly using the
character value as an index into that array, but now that I'm going to
be sorting and changing things, I can no longer use that as a fast way
to encode a character.