Expanded allNodes for Decoding
Currently, the tree that is created on decoding is too large to fit in
the normal 2 * n - 1. I'm assuming that this is just something that
happens with the canonical encoding, but I need to read up on it. For
now, increased to 3 * n size to hold everything so I can continue.